move example client out of src to break import cycles
