+++ plugins/cost-insights/src/index.ts
@@ -15,6 +15,7 @@
  */
 
 export { plugin } from './plugin';
+export * from './ExampleCostInsightsClient';
 export * from './api';
 export * from './components';
 export { useCurrency } from './hooks';
